+++ src/app.module.ts
@@ -12,7 +12,7 @@ import { IndexerModule } from './indexer/indexer.module';
 import { EthereumListenerModule } from './listeners/ethereum-listener/ethereum-listener.module';
 import { SubstrateListenerModule } from './listeners/substrate-listener/substrate-listener.module';
 import { SchedulersModule } from './schedulers/schedulers.module';
-import { keyList, SecretKeyList } from './secrets';
+import { keyList, SecretKeyList } from './common/secrets';
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 require('dotenv').config();
