+++ src/common/common.module.ts
@@ -5,7 +5,7 @@ import {
 import { Module } from '@nestjs/common';
 import { CqrsModule } from '@nestjs/cqrs';
 import { ElasticsearchModule } from '@nestjs/elasticsearch';
-import { keyList, SecretKeyList } from '../secrets';
+import { keyList, SecretKeyList } from '../common/secrets';
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 require('dotenv').config();
