+++ src/common/debio-conversion/debio-conversion.module.ts
@@ -1,7 +1,7 @@
 import { GCloudSecretManagerModule } from '@debionetwork/nestjs-gcloud-secret-manager';
 import { Module } from '@nestjs/common';
 import { DebioConversionService } from './debio-conversion.service';
-import { SecretKeyList } from '../../secrets';
+import { SecretKeyList } from '../../common/secrets';
 
 @Module({
   imports: [
