+++ src/common/substrate/index.ts
@@ -1,7 +1,7 @@
 import { GCloudSecretManagerModule } from '@debionetwork/nestjs-gcloud-secret-manager';
 import { Module } from '@nestjs/common';
 import { SubstrateService } from './substrate.service';
-import { SecretKeyList } from '../../secrets';
+import { SecretKeyList } from '../../common/secrets';
 
 @Module({
   imports: [
