+++ src/schedulers/genetic-analyst-unstaked/unstaked.service.ts
@@ -7,7 +7,7 @@ import {
   retrieveGeneticAnalystUnstakeAmount,
 } from '@debionetwork/polkadot-provider';
 import { GCloudSecretManagerService } from '@debionetwork/nestjs-gcloud-secret-manager';
-import { keyList } from '../../secrets';
+import { keyList } from '../../common/secrets';
 
 @Injectable()
 export class GeneticAnalystUnstakedService implements OnModuleInit {
