+++ src/schedulers/lab-unstake/lab-unstake.service.ts
@@ -7,7 +7,7 @@ import {
   retrieveLabUnstakeAmount,
 } from '@debionetwork/polkadot-provider';
 import { GCloudSecretManagerService } from '@debionetwork/nestjs-gcloud-secret-manager';
-import { keyList } from '../../secrets';
+import { keyList } from '../../common/secrets';
 
 @Injectable()
 export class LabUnstakedService implements OnModuleInit {
