+++ src/schedulers/mailer/mailer.service.ts
@@ -9,7 +9,7 @@ import {
 } from '../../common';
 import { queryLabById } from '@debionetwork/polkadot-provider';
 import { GCloudSecretManagerService } from '@debionetwork/nestjs-gcloud-secret-manager';
-import { keyList } from '../../secrets';
+import { keyList } from '../../common/secrets';
 
 @Injectable()
 export class MailerService {
