+++ src/schedulers/unstaked/unstaked.service.ts
@@ -7,7 +7,7 @@ import {
   retrieveUnstakedAmount,
 } from '@debionetwork/polkadot-provider';
 import { GCloudSecretManagerService } from '@debionetwork/nestjs-gcloud-secret-manager';
-import { keyList } from '../../secrets';
+import { keyList } from '../../common/secrets';
 
 @Injectable()
 export class UnstakedService implements OnModuleInit {
